refactor: added pvcs fluent interface implementation
